remove unnecessary cast
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk@3435 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ jasperreports/src/net/sf/jasperreports/engine/export/JRXhtmlExporter.java
@@ -2083,7 +2083,7 @@ public class JRXhtmlExporter extends JRAbstractExporter
 
 	public String toSizeUnit(int size)
 	{
-		return String.valueOf((int)(zoom * (float)size)) + sizeUnit;
+		return String.valueOf((int)(zoom * size)) + sizeUnit;
 	}
